Code cleanup
Prepared a build process based on the creation of a standalone library.

Cleaned up library code, separating implementation and headers.
Splitted the `_start` symbol section to a separated assembly file.

Fixed the implementation of a basic allocator.

Added documentation on all methods in the headers.
Added basic readme and a license.